requirements.txt
face_recognition @ git+https://github.com/thetoby9944/face_recognition
faiss_cpu==1.10.0
fastapi==0.116.0
numpy==2.3.1
Pillow==11.3.0
pydantic==2.11.7
pymongo==4.13.2
python-dotenv==1.1.1
Requests==2.32.4
schedule==1.2.2
streamlit==1.42.0
